Add codeclimate/eslint integration (#13)
* #11 Refactor following node / npm naming convention

* Encrypt slack token

* Refactor lambdas

* Refactor following node / npm naming convention

* Add codeclimate/eslint integration

* Update gitignore

* Remove tests

* Fix eslint issues

* Renamed to lower case

* Update from dev

* Remove package.json from project root folder

* Change E2E_TESTING to none

* Add unit tests

* Add log to show package.json

* Add log to show package.json

* Check generated tests

* Check issue

* Check if package.json exists

* Check if package.json exists

* Add logging to GitDiffWalker

* Fix logging to GitDiffWalker

* Update gitiignore

* Add logging to preinstall

* Add logging to preinstall

* Copy package.json

* Refactor lambdas && add new unit tests

* Update base on latest skeleton

* Undo changes in post-deply hook

+++ bin/test/install.sh
@@ -18,35 +18,60 @@ else
   __IS_CONCURRENT_SCRIPT=${1}
 fi
 
+#####################################
+### Add logging for imported vars ###
+#####################################
+if [ "$TRAVIS" == "true" ] && [ -e "${__VARS_FILE_PATH}" ]; then
+  head -n 20 "${__VARS_FILE_PATH}"
+fi
+
 if [ "$__IS_CONCURRENT_SCRIPT" == "$__NONE" ] || [ "$__IS_CONCURRENT_SCRIPT" == "$__BACKEND" ]; then
 
-  # To disable interactive user interaction like prompts in terminal (an default value is always chosen)
-  export DEEP_NO_INTERACTION=1
-
-  #install also frontend for repos with e2e enabled
+  ##################################################
+  ### install frontend deps where e2e is enabled ###
+  ##################################################
   if [ "$__IS_CONCURRENT_SCRIPT" != "$__NONE" ] && \
-   ([ "${__E2E_WITH_PUBLIC_REPO}" == "${E2E_TESTING}" ] || \
-   ([ "${__E2E_WITH_PRIVATE_REPO}" == "${E2E_TESTING}" ] && [ "${TRAVIS_BRANCH}" == 'stage' ])); then
-    subpath_run_cmd "${__SRC_PATH}" "$__CMD" "$__CMD" "${__IS_CONCURRENT_SCRIPT}"
+   ( ([ "${__E2E_WITH_PUBLIC_REPO}" == "${E2E_TESTING}" ]) || \
+   ([ "${__E2E_WITH_PRIVATE_REPO}" == "${E2E_TESTING}" ] && [ "${CI_FULL}" == "true" ]) ); then
+    subpath_run_cmd "${__SRC_PATH}" "$__CMD" "$__CMD" "$__FRONTEND"
   fi
 
   echo "Start initializing backend"
 
+  ###########################################################################
+  ### Check if deeploy.example.json exists, otherwise exit with error msg ###
+  ###########################################################################
+  if [ ! -f ${__SRC_PATH}deeploy.example.json ]; then
+    echo "File deeploy.example.json doesn't exist. Please create it and repeat"
+    exit 1
+  fi
+
   cp ${__SRC_PATH}deeploy.example.json ${__SRC_PATH}deeploy.json
 
-  if [ "${TRAVIS}" == "true" ] && [ "$BACKEND_MICROAPP_IDENTIFIERS" == "none" ]; then
-    echo "Skipping initializing backend, becuase no changes in backend"
-  elif [ "$BACKEND_MICROAPP_IDENTIFIERS" == "$__NONE" ] || [ -z "$BACKEND_MICROAPP_IDENTIFIERS" ] || \
-    ( ([ "BACKEND_MICROAPP_PATHS" != "$__NONE" ] || [ "FRONTEND_MICROAPP_PATHS" != "$__NONE" ]) && \
-    ([ "${__E2E_WITH_PUBLIC_REPO}" == "${E2E_TESTING}" ] || ([ "${__E2E_WITH_PRIVATE_REPO}" == "${E2E_TESTING}" ] && \
-    [ "${TRAVIS_BRANCH}" == 'stage' ]))); then
+  if [ "${TRAVIS}" == "true" ] && [ "$BACKEND_MICROAPP_IDENTIFIERS" == "none" ] && \
+    ([ "${FRONTEND_MICROAPP_PATHS}" == "$__NONE" ] || [ "${__E2E_WITH_PUBLIC_REPO}" != "${E2E_TESTING}" ]); then
 
+    ###########################################################################
+    ### Skip initializing backend if no changes in backend or running in CI ###
+    ###########################################################################
+    echo "Skipping initializing backend, becuase no changes in backend"
+  elif ( ([ "${BACKEND_MICROAPP_PATHS}" != "$__NONE" ] || [ "${FRONTEND_MICROAPP_PATHS}" != "$__NONE" ]) && \
+    ([ "${__E2E_WITH_PUBLIC_REPO}" == "${E2E_TESTING}" ] || ([ "${__E2E_WITH_PRIVATE_REPO}" == "${E2E_TESTING}" ] && \
+    [ "${CI_FULL}" == "true" ]))); then
+
+    #################################################################
+    ### Fully initializing backend for e2e test or forced locally ###
+    ##########################################################
     echo "Fully initializing backend"
 
     cd ${__SRC_PATH} && deepify compile dev "${__SRC_PATH}"
   else
-    echo "Partially initializing backend: ${BACKEND_MICROAPP_IDENTIFIERS}"
-    cd ${__SRC_PATH} && deepify compile dev -m "${BACKEND_MICROAPP_IDENTIFIERS}"
+
+    ################################################################
+    ### Partially initializing backend for specified identifiers ###
+    ################################################################
+    echo "Partially initializing backend: ${BACKEND_MICROAPP_PATHS}"
+    cd ${__SRC_PATH} && deepify compile dev -m "${BACKEND_MICROAPP_PATHS}"
   fi
 
 fi
